Refactor adding permissions and leader notifications

+++ app/models/person.rb
@@ -225,18 +225,20 @@ class Person < ActiveRecord::Base
     org_permissions = organizational_permissions.where("organizational_permissions.organization_id = ? AND organizational_permissions.archive_date IS NULL AND organizational_permissions.deleted_at IS NULL", org_id)
     return org_permissions.first.try(:permission) unless org_permissions.count > 1
 
-    admin = org_permissions.where("organizational_permissions.permission_id = ?", Permission::ADMIN_ID)
-    if admin.present?
-      org_permission = admin.first
-      org_permissions_with_archived.where("organizational_permissions.id <> ?", org_permission.id).destroy_all
-      return org_permission.try(:permission)
-    end
+    if self.email.present?
+      admin = org_permissions.where("organizational_permissions.permission_id = ?", Permission::ADMIN_ID)
+      if admin.present?
+        org_permission = admin.first
+        org_permissions_with_archived.where("organizational_permissions.id <> ?", org_permission.id).destroy_all
+        return org_permission.try(:permission)
+      end
 
-    user = org_permissions.where("organizational_permissions.permission_id = ?", Permission::USER_ID)
-    if user.present?
-      org_permission = user.first
-      org_permissions_with_archived.where("organizational_permissions.id <> ?", org_permission.id).destroy_all
-      return org_permission.try(:permission)
+      user = org_permissions.where("organizational_permissions.permission_id = ?", Permission::USER_ID)
+      if user.present?
+        org_permission = user.first
+        org_permissions_with_archived.where("organizational_permissions.id <> ?", org_permission.id).destroy_all
+        return org_permission.try(:permission)
+      end
     end
 
     contact = org_permissions.where("organizational_permissions.permission_id = ?", Permission::NO_PERMISSIONS_ID)
